Change the return value of Nix::Platform::audioHardwareSampleRate from double to float.
So it stay in sync with the type used by WebCore.

Reviewed by: Marcelo Lira

+++ Source/Platform/nix/public/Platform.h
@@ -49,7 +49,7 @@ public:
     WEBKIT_EXPORT static Platform* current();
 
     // Audio --------------------------------------------------------------
-    virtual double audioHardwareSampleRate() { return 0; }
+    virtual float audioHardwareSampleRate() { return 0; }
     virtual size_t audioHardwareBufferSize() { return 0; }
     virtual unsigned audioHardwareOutputChannels() { return 0; }
